backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
python-magic==0.4.27
